+++ crates/protocols/hpke/libcrux_provider/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- [#98](https://github.com/cryspen/hpke-rs/pull/98): add support for AES-GCM to the provider
+
 ## 0.3.0 - 2025-07-01
 
 * initial release